Move execution code in a separate package
This is part of a bigger refactor around how execution code is organized.
