+++ bionic/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Contains Bionic core logic to execute tasks and their dependencies.
+"""
